Rename the dist folder 'plugins' (#444)
