--- index.js
@@ -1,2 +1,0 @@
-// stub: this is package.json's "main"
-// truffle require.resolve's this in order to check plugin is installed
