+++ test/integration/projects/libraries/buidler.config.js
@@ -1,5 +1,5 @@
 const { loadPluginFile } = require("@nomiclabs/buidler/plugins-testing");
-loadPluginFile(__dirname + "/../dist/buidler.plugin");
+loadPluginFile(__dirname + "/../plugins/buidler.plugin");
 usePlugin("@nomiclabs/buidler-truffle5");
 
 module.exports={
